+++ release/ui/cards/FlipCard/FlipCard.js
@@ -24,7 +24,12 @@ function FlipCard(_a = {}) {
             revertAnimation && styles.revert,
             styles[animation],
             dom.use(flip) && styles.flip,
-        ]},children:[{type:Card.Card,props:{flex:true,...props,onclick:handleClick,class:() => styles.front},children:[children]},{type:Card.Card,props:{...props,onclick:handleClick,class:() => styles.back},children:[backside]}]});
+        ]},children:[{type:Card.Card,props:{flex:true,...props,onclick:handleClick,class:{
+            root: () => styles.front,
+            show: () => styles.show,
+            shown: () => styles.shown,
+            hide: () => styles.hide,
+        }},children:[children]},{type:Card.Card,props:{...props,onclick:handleClick,class:() => styles.back},children:[backside]}]});
 }
 
 exports.FlipCard = FlipCard;