+++ release/ui/popups/Modal/Modal.js
@@ -9,10 +9,9 @@ var classes = require('html-classes');
 var watchState = require('watch-state');
 require('../../../utils/index.js');
 require('../../buttons/index.js');
-require('../../icons/index.js');
 require('../../layout/index.js');
 var Modal$1 = require('./Modal.scss.js');
-var Icon = require('../../icons/Icon/Icon.js');
+var CloseButton = require('../../buttons/CloseButton/CloseButton.js');
 var setOverflow = require('../../../utils/setOverflow/setOverflow.js');
 var Flex = require('../../layout/Flex/Flex.js');
 var Button = require('../../buttons/Button/Button.js');
@@ -23,6 +22,7 @@ var classes__default = /*#__PURE__*/_interopDefaultLegacy(classes);
 
 const useStyles = dom.style(Modal$1["default"]);
 let modalsCount = 0;
+const defaultCloseButton = {type:CloseButton.CloseButton,props:{offset:8,padding:8,size:16}};
 function Modal(_a = {}) {
     var { buttons, width, height, style, headButtons = ['close'], buttonProps = {}, onclosed, onclose, onshow, onmousedown } = _a, props = tslib.__rest(_a, ["buttons", "width", "height", "style", "headButtons", "buttonProps", "onclosed", "onclose", "onshow", "onmousedown"]);
     const styles = useStyles();
@@ -33,7 +33,7 @@ function Modal(_a = {}) {
     const headButtonsLength = headButtons === null || headButtons === void 0 ? void 0 : headButtons.length;
     const buttonsLength = buttons === null || buttons === void 0 ? void 0 : buttons.length;
     if (!('button-close' in slots)) {
-        slots['button-close'] = {type:Icon.Icon,props:{icon:'cross'}};
+        slots['button-close'] = defaultCloseButton;
     }
     setTimeout(() => {
         show.value = true;
@@ -67,7 +67,13 @@ function Modal(_a = {}) {
                 ((_a = hidden.value) === null || _a === void 0 ? void 0 : _a.value) && styles.hide,
                 !title && !subTitle && styles.noTitle,
             ]);
-        }},children:[{type:'show',props:{state:title || subTitle || headButtonsLength},children:[{type:Flex.Flex,props:{padding:24,element:'header',vertical:true,align:'stretch',class:() => styles.header},children:[{type:'show',props:{state:title || headButtonsLength},children:[{type:Flex.Flex,props:{flex:true},children:[{type:Flex.Flex,props:{flex:true},children:[title || null]},{type:'show',props:{state:headButtonsLength},children:[{type:'div',props:{class:() => styles.headButtons},children:[headButtons.map(id => ({type:'button',props:{...buttonProps[id],'data-button':id,onclick:() => handleClose(id),class:() => styles.headButton},children:[slots[`button-${id}`] || id]}))]}]}]}]},{type:'show',props:{state:subTitle},children:[{type:'div',props:{class:() => styles.subTitle},children:[subTitle]}]}]}]},content && ({type:'div',props:{class:() => styles.content},children:[content]}),' ',children,{type:'show',props:{state:buttonsLength},children:[{type:Flex.Flex,props:{reverse:true,justify:'center',wrap:true,padding:16,gap:16},children:[buttons === null || buttons === void 0 ? void 0 : buttons.map((id, index) => ({type:Button.Button,props:{flex:true,view:index ? 'secondary' : 'primary',...buttonProps[id],'data-button':id,onclick:() => handleClose(id)},children:[slots[`button-${id}`] || id]}))]}]}]}]});
+        }},children:[{type:'show',props:{state:title || subTitle || headButtonsLength},children:[{type:Flex.Flex,props:{padding:24,element:'header',vertical:true,align:'stretch',class:() => styles.header},children:[{type:'show',props:{state:title || headButtonsLength},children:[{type:Flex.Flex,props:{flex:true},children:[{type:Flex.Flex,props:{flex:true},children:[title || null]},{type:'show',props:{state:headButtonsLength},children:[{type:'div',props:{class:() => styles.headButtons},children:[headButtons.map(id => ({
+            type: id === 'close' ? 'span' : 'button',
+            props: Object.assign(Object.assign({}, buttonProps[id]), { 'data-button': id, onclick: () => handleClose(id), class: () => styles.headButton }),
+            children: [
+                slots[`button-${id}`] || id,
+            ],
+        }))]}]}]}]},{type:'show',props:{state:subTitle},children:[{type:'div',props:{class:() => styles.subTitle},children:[subTitle]}]}]}]},content && ({type:'div',props:{class:() => styles.content},children:[content]}),' ',children,{type:'show',props:{state:buttonsLength},children:[{type:Flex.Flex,props:{reverse:true,justify:'center',wrap:true,padding:16,gap:16},children:[buttons === null || buttons === void 0 ? void 0 : buttons.map((id, index) => ({type:Button.Button,props:{flex:true,view:index ? 'secondary' : 'primary',...buttonProps[id],'data-button':id,onclick:() => handleClose(id)},children:[slots[`button-${id}`] || id]}))]}]}]}]});
 }
 
 exports.Modal = Modal;